Add style declarations for the recipe pages
Add style to the h1 to help it stick out more, by using a red background
color. Change the image size to a specific width and height to auto so
the pages will be similar across recipes no matter the initial size of
their images.
Center text on most text content elements, but not the ingredients or
cooking steps.
Underline the ingredients and steps to add some non semantic importance
to them.